requirements.txt
aiohttp==3.7.3
astroid==2.3.0
async-timeout==3.0.1
attrs==20.3.0
awscli==1.18.217
backcall==0.2.0
backoff==1.10.0
beautifulsoup4==4.9.3
bidict==0.21.2
botocore==1.19.57
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Django==2.0.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.1
git-remote-codecommit==1.15.1
idna==2.10
idna-ssl==1.1.0
ikp3db==1.4.1
importlib-metadata==3.4.0
Inflector==2.0.12
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.5.2
lyricsgenius==3.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
parso==0.1.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.11
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
Pygments==2.7.4
pylint==2.4.4
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
spotify==0.10.2
spotipy==2.16.1
SQLAlchemy==1.3.23
stevedore==3.3.0
traitlets==4.3.3
typed-ast==1.2.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
zipp==3.4.0
